[feat] add NoResult in PlantList (change lottie embedding)
